requirements.txt
tkinter
watchdog
python-vlc
Pillow
RPi.GPIO
